+++ packages/dataviews/src/stories/dataform.story.tsx
@@ -1384,7 +1384,7 @@ const LayoutCardComponent = ( {
 			isVisible: ( item ) => item.displayPayments,
 			render: ( { item } ) => {
 				return (
-					<p>
+					<p style={ { margin: 0 } }>
 						The customer has made a total of { item.totalOrders }{ ' ' }
 						orders, amounting to { item.totalRevenue } dollars. The
 						average order value is { item.averageOrderValue }{ ' ' }
@@ -1440,17 +1440,13 @@ const LayoutCardComponent = ( {
 
 	const form: Form = useMemo(
 		() => ( {
-			layout: getLayoutFromStoryArgs( {
-				type: 'card',
-				withHeader,
-			} ),
 			fields: [
 				{
 					id: 'customerCard',
 					layout: {
 						type: 'card',
 						summary: 'plan-summary',
-						isCollapsible,
+						withHeader,
 					},
 					label: 'Customer',
 					description: